Adds the Stack implementation for js
This stack implementation is designed to either return a stack
or add stack functionality to an object that is passed in to it.

Stacks are pretty simple so the implementation is pretty simple.
Lots of ES2015 in side to make it terse because there is really
nothing complicated about the code.