Adds in implementation for remove, removeAt, and indexOf
This commit adds the above methods to the linked list implementation.
remove is essentially just a composite of the other two.

Expands tests to cover the new additions.

I started moving some of the methods out of the module reveal to
allow for compositing them for things like .remove. I may continue
this in the future as other composite methods are created.